third_party/WebKit/LayoutTests/editing/selection/DOMSelection-DocumentType-expected.txt

Test to check if setBaseAndExtent guard node with null owner document (Bug 31680)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.extend(docType, 0) threw exception InvalidStateError: Failed to execute 'extend' on 'Selection': This Selection object doesn't have any Ranges..
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.containsNode(docType) is false
PASS successfullyParsed is true

TEST COMPLETE
